+++ joint/autocorrelation_test.go
@@ -14,7 +14,7 @@ import (
 
 func TestNewAutocorrelation(t *testing.T) {
 	autocorrelation := NewAutocorrelation(1, 3)
-	assert.Equal(t, uint64(1), autocorrelation.lag)
+	assert.Equal(t, 1, autocorrelation.lag)
 	assert.Equal(t, NewCorrelation(3), autocorrelation.correlation)
 }
